requirements.txt
pandas~=2.2.2
numpy~=1.26.1
matplotlib~=3.8.4
pillow~=10.3.0
opencv-python~=4.10.0.84
mediapipe~=0.10.14
openpyxl~=3.1.5

